+++ changelog.md
@@ -8,12 +8,11 @@ Uiua is not yet stable.
 ### Language
 - Add the experimental [`cascade` `⪾`](https://uiua.org/docs/cascade) modifier, which captures some common [`fork` `⊃`](https://uiua.org/docs/fork) patterns
 - Add experimental hashmap functions, which operate on a box array as if it is a hashmap
+  - [`map`](https://uiua.org/docs/map)
   - [`has`](https://uiua.org/docs/has)
   - [`get`](https://uiua.org/docs/get)
   - [`insert`](https://uiua.org/docs/insert)
   - [`remove`](https://uiua.org/docs/remove)
-  - [`keys`](https://uiua.org/docs/keys)
-  - [`values`](https://uiua.org/docs/values)
 ### Interpreter
 - The internal byte array type is now used in more places, which should improve performance a bit
 - Bug and crash fixes